Account: share error handling btwn brave://account and brave://settings
