+++ browser/resources/settings/sources.gni
@@ -223,6 +223,10 @@ brave_settings_mojo_files = [
   "$root_gen_dir/brave/components/brave_account/mojom/reset_password.mojom-webui.ts",
   "$root_gen_dir/brave/components/brave_origin/mojom/brave_origin_settings.mojom-webui.ts",
 
+  # Shared with the Brave Account UI; lands at preprocessed/
+  # brave_account_shared.ts.
+  "//brave/components/brave_account/resources/brave_account_shared.ts",
+
   # Add generated strings as mojo files. Currently, this is the only way to
   # include them in build_webui() target.
   "$root_gen_dir/brave/grit/brave_generated_resources_webui_strings.ts",
