+++ components/brave_account/resources/brave_account_sign_in_dialog.ts
@@ -11,8 +11,8 @@ import {
 } from './brave_account_browser_proxy.js'
 import { getCss } from './brave_account_sign_in_dialog.css.js'
 import { getHtml } from './brave_account_sign_in_dialog.html.js'
-import { LoginClientErrorCode, LoginError } from './login.mojom-webui.js'
-import { showError } from './brave_account_common.js'
+import { LoginClientErrorCode } from './login.mojom-webui.js'
+import { showError } from './brave_account_shared.js'
 
 import {
   invalidLoginError,
@@ -73,27 +73,11 @@ export class BraveAccountSignInDialogElement extends CrLitElement {
         clientMac,
       )
     } catch (e) {
-      let error: LoginError
-
-      if (e && typeof e === 'object') {
-        error = e as LoginError
-      } else if (typeof e === 'string') {
-        error = {
-          clientError: {
-            errorCode:
-              e === invalidLoginError()
-                ? LoginClientErrorCode.kInvalidLoginError
-                : LoginClientErrorCode.kOpaqueError,
-          },
-        }
-      } else {
-        console.error('Unexpected error:', e)
-        error = {
-          clientError: { errorCode: LoginClientErrorCode.kUnexpected },
-        }
-      }
-
-      showError({ kind: 'login', details: error })
+      showError('login', e, {
+        opaqueErrors: {
+          [invalidLoginError()]: LoginClientErrorCode.kInvalidLoginError,
+        },
+      })
     }
 
     this.isSubmitting = false
